feat: refactor validation
