+++ modules/validation/duplicates.js
@@ -1,0 +1,3 @@
+const getDuplicates = (array) => new Set(array).size !== array.length;
+
+module.exports = getDuplicates;
